requirements.txt
Django>=2.1.2
djangorestframework
drf-nested-routers
psycopg2
